Adding createDefaults method
